Organize auxiliary scripts
