+++ scripts/setup_master.sh
